requirements.txt
Flask==3.0.3
protobuf==5.29.1
grpcio-tools==1.68.1
tensorflow==2.16.2
tensorflow-intel==2.16.2
